refactor: simplify and organize imports
Simplify import paths ending with `../src` or `../index` by leaving out
the last part which is not needed.

Run VSCode's "Organize Import" command in the modified files to keep
import statements organized in a deterministic way.
